feat(OmokGameActivity): OmokGameController에 있던 코드를 다시 OmokGameActivity로 가져옴
